Use Creative Commons Share Alike By Attribution
The GNU FDL makes alot sense for the Clemson ACM Hackpack because the
technical manuals that are often referenced for C++ programming are also
under that license.  However after further exploration, most Cyber
Security content appears to be under the Creative Commons License
Attribution-ShareAlike  license. The two licenses are incompatible due
the presence of invariant sections and cover pages in the GNU FDL.  As
such we need to use the license that makes the most sense for the
content that we are attempting to educate.

However, we should still use the GPL v3 for the license for all shell
scripts and programs that we write because it provides the
necessary freedoms that will allow the hackpack and other similar
projects to grow.