Fix a bug in the firewalld configuration example
Bash does not preform variable interpolation on single quoted strings.
This prevented the `$addr` in the script from being correctly expanded.
Which in turn would cause the script to fail with an invalid syntax
error.

Also fixed was an error where the word 'address' was left out of the
source specification in the rich rules.